Fix new run script

+++ src/main/scala/com/alexpyc/EncryptedJoin.scala
@@ -18,7 +18,7 @@ object EncryptedJoin {
       .getOrCreate()
     Utils.initSQLContext(spark.sqlContext)
 
-    val numPartitions = args(0)
+    val numPartitions = args(0).toInt
     val sizeLevel = args(1).toInt
 
     JoinCost.run(spark, Encrypted, (math.pow(2, sizeLevel) * 125).toInt.toString, numPartitions)
