+++ src/main/scala/com/alexpyc/EncryptedPR.scala
@@ -18,7 +18,7 @@ object EncryptedPR {
       .getOrCreate()
     Utils.initSQLContext(spark.sqlContext)
 
-    val numPartitions = args(0)
+    val numPartitions = args(0).toInt
     val sizeLevel = args(1).toInt
 
     PageRank.run(spark, Encrypted, math.pow(2, sizeLevel).toInt.toString, numPartitions)
