+++ src/main/scala/com/alexpyc/EncryptedQ2.scala
@@ -18,7 +18,7 @@ object EncryptedQ2 {
       .getOrCreate()
     Utils.initSQLContext(spark.sqlContext)
 
-    val numPartitions = args(0)
+    val numPartitions = args(0).toInt
     val sizeLevel = args(1).toInt
 
     val size = math.pow(10, i).toInt.toString
