+++ src/main/scala/com/alexpyc/InsecureJoin.scala
@@ -18,7 +18,7 @@ object InsecureJoin {
       .getOrCreate()
     Utils.initSQLContext(spark.sqlContext)
 
-    val numPartitions = args(0)
+    val numPartitions = args(0).toInt
     val sizeLevel = args(1).toInt
 
     JoinCost.run(spark, Insecure, (math.pow(2, sizeLevel) * 125).toInt.toString, numPartitions)
